Rent model and SearchRentRequest Tests added, minor changes to Rent,R…

+++ car-rental-office-core/src/main/java/hu/uni/miskolc/iit/service/RentManagementService.java
@@ -17,7 +17,7 @@ public interface RentManagementService {
      * @param id
      * @return Rent object
      * */
-    Rent getRentById(long id);
+    Rent getRentById(int id);
 
     /**
      * A szűrési paramétereknek megfelelő kölcsönöket adja vissza
